requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.2.0
pip==23.3.1
setuptools==68.2.2
Werkzeug==3.0.1
wheel==0.41.2